Update index.md
"We organize data in spreadsheets in the ways that we as humans want to work with the data, but computers require that data be organized in particular ways." very clunky long sentence.  I have edited to make it two sentences, as well as changed wording slightly so it reads better. Now reads "Typically we organize data in spreadsheets in ways that we as humans want to work with the data. However 
computers require data to be organized in particular ways." Part of carpentry instructor checkout