Update 02-common-mistakes.md
The current figure of multiple tables may be challenging to comprehend as it does not properly match the description that follows it.

What we have:
![multiple tables](../fig/multiple-tables-example.png)
In the example above, the computer will see (for example) row 24 and assume that all columns A-J 
refer to the same sample. This row actually represents two distinct samples 
(information about livestock for informant 1 and information about plots for informant 2)


The improvements it needs:
it refers to row 24 and columns A to J 
but the figure does not have any row header or column header.
So, learners get confused about which row and column of multiple-tables-example.png
the description is referring to.


Appropriate changes that have been done:
I have uploaded a 2nd file and updated the link on the page.
![multiple tables](../fig/multiple-tables-example2.png)
which improves the SAFI_messy Screenshot with Row and Column Headings while having the row 24 marked active.