Remove confusing speaker notes from lesson content
The "Introduce the Data" section heading and first sentence after it is a call-to-action for the trainer rather than for learners to read. All other content in the lesson is directed to the learner. Personally I find this is confusing and I lose the flow when trying to present thie material when I encounter these speaker notes.  The proposal is to rename this Pinned section  to "Workshop Data" and delete the first sentence. This means there are no longer any speaker notes in the content of this lesson and hopefully it will flow more easily, for both the trainer and the learner.